Appveyor: Remove "package installation workaround" (#2464)
The `pkgbuild` bug is fixed and we can use the latest version of *tic* again. 

To get a green build, we need to set a `Makevars.win` file to meet the C++ standards on Windows.

I faced some errors with the `digest` installation in some tries and installing it explicitly helped. Don't wanna run more builds with and without cache, maybe the issue also came from a faulty cache. 

*tic* hasn't yet changed the arguments of `step_rcmdcheck()`. This only happened in my fork so we need to use the "old" `warnings_are_errors` and `notes_are_errors` arguments again.